repository/finance-api/pyproject.toml
[project]
name = "fitech-finance"
version = "0.1.0"
description = ""
authors = [
    {name = "Mateus Portela"}
]
requires-python = "^3.13"
dependencies = [
    "fastapi[standard] (>=0.136.1,<0.137.0)",
    "tortoise-orm[aiomysql] (>=0.21,<1.0)",
    "aerich (>=0.9.2,<0.10.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pydantic[email] (>=2.13.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
